requirements.txt.txt
pandas
openpyxl
lxml
pdfminer.six
openai-whisper
fasttext
faiss-cpu
python-dotenv
SQLAlchemy
nltk
tqdm
numpy
scikit-learn